fix(api): consume host-local MicroRegistry aliases
