+++ firecrab-api/src/handlers.rs
@@ -3,6 +3,10 @@ pub mod builder_vm;
 pub mod console;
 pub mod console_sentinel;
 pub mod images;
+mod images_core;
+#[cfg(test)]
+#[allow(clippy::err_expect)]
+mod images_local_tests;
 pub mod kernels;
 pub mod micro_networks;
 pub mod micro_storages;
